requirements.txt
google-cloud-storage==2.13.0
requests==2.31.0 